Adds some base classes for a fun little idea for an app
App to allow people to connect to private google maps list of local resources for foragers. the vision is a place where people can report certain types of resources and foragers can share resources. humans are not good at this and this app could also easily be weaponized by the contents of the map and should never go live. but I will be able to run a local list of edible plants and funguses and pass it around to those who would use it wisely.